Formation-in-screen routine fixed and other improvements.
Cought a weird bug during formation-in-screen check. Due to the pools, the
enemies being recycled were out of the screen. Even calling reset with 0,0
as X,Y coordinates is not enough to update the bounding box calculation to
the proper vaues. Now forcing an updateTransform() when resetting the enemy.

Also added handling formation repetition.